[move] `node()` の実装を Builder+Generator ファイルに移動しました.
